Move tests from unit to e2e.
